+++ src/Config.php
@@ -1130,7 +1130,11 @@ class Config
             ],
             'maintenance_mode'            => false,
             'headers'                     => [
-                'x_frame_options'     => true,
+                'x_frame_options' => true,
+            ],
+            'htmlcleaner'                 => [
+                'allowed_tags'       => explode(',', 'div,span,p,br,hr,s,u,strong,em,i,b,li,ul,ol,mark,blockquote,pre,code,tt,h1,h2,h3,h4,h5,h6,dd,dl,dh,table,tbody,thead,tfoot,th,td,tr,a,img,address,abbr'),
+                'allowed_attributes' => explode(',', 'id,class,style,name,value,href,src'),
             ],
         ];
     }