+++ src/Storage/Field/Type/HtmlType.php
@@ -1,6 +1,10 @@
 <?php
+
 namespace Bolt\Storage\Field\Type;
 
+use Bolt\Storage\Field\Sanitiser\SanitiserAwareInterface;
+use Bolt\Storage\Field\Sanitiser\SanitiserAwareTrait;
+use Bolt\Storage\Field\Sanitiser\WysiwygAwareInterface;
 use Doctrine\DBAL\Types\Type;
 
 /**
@@ -9,8 +13,10 @@ use Doctrine\DBAL\Types\Type;
  *
  * @author Ross Riley <riley.ross@gmail.com>
  */
-class HtmlType extends FieldTypeBase
+class HtmlType extends FieldTypeBase implements SanitiserAwareInterface, WysiwygAwareInterface
 {
+    use SanitiserAwareTrait;
+
     /**
      * {@inheritdoc}
      */